requirements.txt
# Core Django
Django==4.2.0
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.11

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.9.0

# Swagger/OpenAPI
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1

# Redis & Caching
redis==5.0.7
django-redis==5.4.0

# Celery
celery==5.4.0

# Utilities
requests==2.32.3
python-dotenv==1.0.0
memory_profiler==0.61.0
colorlog==6.8.0
qrcode[pil]==7.4.2
django-filter==23.5

# Telegram Bot
python-telegram-bot==21.0
qrcode[pil]==7.4.2